// src/api/models/ClassesPost201Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend API
 * API documentation for the backend service
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SchoolClass } from './SchoolClass';
import {
    SchoolClassFromJSON,
    SchoolClassFromJSONTyped,
    SchoolClassToJSON,
    SchoolClassToJSONTyped,
} from './SchoolClass';

/**
 * 
 * @export
 * @interface ClassesPost201Response
 */
export interface ClassesPost201Response {
    /**
     * 
     * @type {number}
     * @memberof ClassesPost201Response
     */
    code?: number;
    /**
     * 
     * @type {string}
     * @memberof ClassesPost201Response
     */
    message?: string;
    /**
     * 
     * @type {SchoolClass}
     * @memberof ClassesPost201Response
     */
    _class?: SchoolClass;
}

/**
 * Check if a given object implements the ClassesPost201Response interface.
 */
export function instanceOfClassesPost201Response(value: object): value is ClassesPost201Response {
    return true;
}

export function ClassesPost201ResponseFromJSON(json: any): ClassesPost201Response {
    return ClassesPost201ResponseFromJSONTyped(json, false);
}

export function ClassesPost201ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ClassesPost201Response {
    if (json == null) {
        return json;
    }
    return {
        
        'code': json['code'] == null ? undefined : json['code'],
        'message': json['message'] == null ? undefined : json['message'],
        '_class': json['class'] == null ? undefined : SchoolClassFromJSON(json['class']),
    };
}

export function ClassesPost201ResponseToJSON(json: any): ClassesPost201Response {
    return ClassesPost201ResponseToJSONTyped(json, false);
}

export function ClassesPost201ResponseToJSONTyped(value?: ClassesPost201Response | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'code': value['code'],
        'message': value['message'],
        'class': SchoolClassToJSON(value['_class']),
    };
}