// src/api/models/Student.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Backend API
 * API documentation for the backend service
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Student
 */
export interface Student {
    /**
     * 
     * @type {number}
     * @memberof Student
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    email?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Student
     */
    userId?: number | null;
    /**
     * 
     * @type {Date}
     * @memberof Student
     */
    createdAt?: Date | null;
}

/**
 * Check if a given object implements the Student interface.
 */
export function instanceOfStudent(value: object): value is Student {
    return true;
}

export function StudentFromJSON(json: any): Student {
    return StudentFromJSONTyped(json, false);
}

export function StudentFromJSONTyped(json: any, ignoreDiscriminator: boolean): Student {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'email': json['email'] == null ? undefined : json['email'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function StudentToJSON(json: any): Student {
    return StudentToJSONTyped(json, false);
}

export function StudentToJSONTyped(value?: Student | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'email': value['email'],
        'userId': value['userId'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    };
}